Personal Learning/02_run_config/pyproject.toml
[project]
name = "02-run-config"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.38.0",
    "openai-agents>=0.3.2",
]
